B and B'
Did B and B' but somthing is wrong with cube[0]
And L and L' is not correct. Just copy pasted a place holder and i finished the comments and changed the names of the turnes to be shorter. Also wrote down the first algorithm. When the turns are finished it should work.