Added new case checking for White Cross, etc.
It has mistakes in algorithm, idk why. I write the algorithm in the old code before and it worked. Then I see we have in the main Branch a new version and I combined my algorithm with the new version, so my code get problems unfortunately

I've updated some things. It is for Daniel

Fixed a few errors :)

Co-authored-by: Reeder Ward