added link to algorithms spreadsheet
* f2l added

* PLL if else added

* added link to algorithms spreadsheet

* checked OLL, PLL, white cross

Checked algs in excel and fixed them in the code

* Fix some F2L algs

Some F2L algs were wrong and now corrected.

* Fix usage of solved and cube in WC and F2L

In the if/ else if statements "cube == solved" was not checking the piece and its position but rather the position and the colour that said position has. I corrected the mistake by switching cube iwht solved in every if/ else if statement. Copied WC and F2L into algorithms.java to be able to use find and replace more easily.

Co-authored-by: ThePunncake <nonamedleight99@gmail.com>