little bit of organizing in the code
* f2l added

* PLL if else added

* added link to algorithms spreadsheet

* checked OLL, PLL, white cross

Checked algs in excel and fixed them in the code

* Fix some F2L algs

Some F2L algs were wrong and now corrected.

* Fix usage of solved and cube in WC and F2L

In the if/ else if statements "cube == solved" was not checking the piece and its position but rather the position and the colour that said position has. I corrected the mistake by switching cube iwht solved in every if/ else if statement. Copied WC and F2L into algorithms.java to be able to use find and replace more easily.

* Add GUI and change little things in solver

Added Ömers GUI code as a new .java file. Implimented Lennarts tips to make the file look cleaner:
-changed some public to private
-added method solve

* add some notes and small things

created solve method added some documentation and checklist

* Minor fixes; like package GUI1 didn't work

Co-authored-by: ThePunncake <nonamedleight99@gmail.com>
Co-authored-by: Reeder Ward II <54606752+reederward1285@users.noreply.github.com>